Move code file to desired folder.
Move 'A.2_3 moves.cpp' to C++ folder.
